refactor #23 파일 구조 변경
