--- lib/core/constants/app_constants.dart
@@ -1,3 +1,0 @@
-class AppConstants {
-  static const String baseUrl = "https://www.debate-seadddson.click/api/v1/";
-}
