--- lib/core/constants/app_theme.dart
@@ -1,32 +1,0 @@
-// import 'package:flutter/material.dart';
-
-// class AppTheme {
-//   static ThemeData get lightTheme {
-//     return ThemeData(
-//       fontFamily: 'Pretendard', // Pretendard를 기본 폰트로 설정
-//       primarySwatch: Colors.blue,
-//       textTheme: const TextTheme(
-//         headline1: TextStyle(
-//           fontSize: 32.0,
-//           fontWeight: FontWeight.bold,
-//           color: Colors.black,
-//         ),
-//         headline2: TextStyle(
-//           fontSize: 24.0,
-//           fontWeight: FontWeight.bold,
-//           color: Colors.black,
-//         ),
-//         bodyText1: TextStyle(
-//           fontSize: 16.0,
-//           fontWeight: FontWeight.normal,
-//           color: Colors.black87,
-//         ),
-//         bodyText2: TextStyle(
-//           fontSize: 14.0,
-//           fontWeight: FontWeight.normal,
-//           color: Colors.black54,
-//         ),
-//       ),
-//     );
-//   }
-// }
