+++ lib/features/auth/presentation/views/login_screen.dart
@@ -1,4 +1,4 @@
-import 'package:debateseason_frontend_v1/features/auth/presentation/view_model/auth_view_model.dart';
+import 'package:debateseason_frontend_v1/features/auth/presentation/view_models/auth_view_model.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
