+++ lib/features/chat/data/datasources/chat_room_api_service.dart
@@ -1,11 +1,12 @@
 import 'dart:convert';
-import 'package:debateseason_frontend_v1/core/constants/app_constants.dart';
+
+import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:http/http.dart' as http;
 
 class ChatRoomApiService {
   Future<Map<String, String>> fetchChatRoomDetails(int chatRoomId) async {
     final response = await http.get(
-      Uri.parse("${AppConstants.baseUrl}/api/v1/$chatRoomId"),
+      Uri.parse("${dotenv.get("BASE_URL")}/api/v1/$chatRoomId"),
     );
 
     if (response.statusCode == 200) {
